+++ src/conditionalLogicService/index.ts
@@ -1,7 +1,0 @@
-export {
-  evaluateConditionalOptionsPredicate,
-  evaluateConditionalPredicate,
-  evaluateConditionalPredicates,
-} from './evaluationService'
-
-export { default as generateFormElementsConditionallyShown } from './generateFormElementsConditionallyShown'