+++ src/index.ts
@@ -6,7 +6,8 @@ import replaceCustomValues, {
 export * as typeCastService from './typeCastService'
 export * as userService from './userService'
 export * as abnService from './abnService'
-export { validatePaymentAmount } from './paymentAmountValidation'
-export { getContentDisposition } from './fileUploadService'
+export * as paymentService from './paymentService'
+export * as schedulingService from './schedulingService'
+export * as fileUploadService from './fileUploadService'
 
 export { replaceCustomValues, getElementSubmissionValue }